% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affliated_cost.R
\name{affliated_cost}
\alias{affliated_cost}
\title{Extract the holding fund cost with other affliated institution}
\usage{
affliated_cost(raw)
}
\arguments{
\item{raw}{Text content from pdf file}
}
\value{
Table of each type of fund cost
}
\description{
Extract the holding fund cost with other affliated institution
}
\examples{
data(pdf_content)
affliated_cost(pdf_content)
}
