% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funds_output.R
\name{funds_output}
\alias{funds_output}
\title{Output the extracted funds statement data}
\usage{
funds_output(code_name_list)
}
\arguments{
\item{code_name_list}{A vector contain code and file name}
}
\value{
Data result
}
\description{
Output the extracted funds statement data
}
\examples{
download.file('http://www.fund.pingan.com/nasfile/1711755883661.pdf', '014645.OF.pdf')
code_name_list<-matrix(c("014645.OF","014645.OF.pdf"),nrow=1,ncol=2)
colnames(code_name_list)<-c('code','filename')
output_dat<-funds_output(data.frame(code_name_list))
}
