% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holding_fee.R
\name{holding_fee}
\alias{holding_fee}
\title{Extract the total holding fund cost}
\usage{
holding_fee(raw)
}
\arguments{
\item{raw}{Text content from pdf file}
}
\value{
3*2 dataframe, where rows corresponding to 3 types of fund cost, columns corresponding to costs' name and data
}
\description{
Extract the total holding fund cost
}
\examples{
data(pdf_content)
holding_fee(pdf_content)
}
