% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statement.R
\name{read_statement}
\alias{read_statement}
\title{Read the text of fund's statement}
\usage{
read_statement(file_name)
}
\arguments{
\item{file_name}{Fund statement's filename or url}
}
\value{
Text string data
}
\description{
Read the text of fund's statement
}
\examples{
example_file_loc<-(system.file("pdfdata", "014645.OF.pdf", package = "cnfund"))
pdf_content<-read_statement(paste(example_file_loc,"//doc//","014645.OF.pdf",sep=''))
}
