% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connnect_table.R
\name{connnect_table}
\alias{connnect_table}
\title{Combine all the statement data together}
\usage{
connnect_table(code, filename)
}
\arguments{
\item{code}{The fund's "OF" code you create}

\item{filename}{The fund's statement file name}
}
\value{
Statement data vector containing all the items
}
\description{
Combine all the statement data together
}
\examples{
connnect_table("'http://www.fund.pingan.com/nasfile/1711755883661.pdf'",sep='')
}
