% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funds_output.R
\name{funds_output}
\alias{funds_output}
\title{Output the extracted funds statement data}
\usage{
funds_output(code_name_list)
}
\arguments{
\item{code_name_list}{A vector contain code and file name}
}
\value{
Data result
}
\description{
Output the extracted funds statement data
}
\examples{
example_file_loc<-(system.file("pdfdata", "014645.OF.pdf", package = "cnfund"))
code_name_list<-c("014645.OF",paste(example_file_loc,"//doc//","014645.OF.pdf", sep=''))
output_dat<-funds_output(code_name_list)
}
