% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statement.R
\name{read_statement}
\alias{read_statement}
\title{Read the text of fund's statement}
\usage{
read_statement(file_name)
}
\arguments{
\item{file_name}{Fund statement's filename or url}
}
\value{
Text string data
}
\description{
Read the text of fund's statement
}
\examples{
download.file('http://www.fund.pingan.com/nasfile/1711755883661.pdf', '014645.OF.pdf')
pdf_content<-read_statement("014645.OF.pdf")
}
